feat: fetch backend api (login, register)

+++ driver_app/lib/Data/models/requests/create_driver_request.dart
@@ -10,7 +10,7 @@ class CreateDriverRequestBody {
   bool Gender;
   String Address;
   double AverageRate;
-  double NumberOfRate;
+  int NumberOfRate;
   int NumberOfTrip;
   String? AccountId;
